Add files via upload
This code is all about the binary search of the array.
Binary Search: Search a sorted array by repeatedly dividing the search interval in half. Begin with an interval covering the whole array. If the value of the search key is less than the item in the middle of the interval, narrow the interval to the lower half. Otherwise, narrow it to the upper half. Repeatedly check until the value is found or the interval is empty.
It is better explained through code go and analyze it there are two approaches to do binary search one is through recursive and the other is using while loop and if-else if ladder.