Add files via upload
This code stores the data of the students using structures and dynamic memory allocation.
Structures: A structure is a user-defined data type in C. A structure creates a data type that can be used to group items of possibly different types into a single type.
Dynamic Memory Allocation: Dynamic Memory Allocation can be defined as a procedure in which the size of a data structure (like Array) is changed during the runtime.
C provides some functions to achieve these tasks. There are 4 library functions provided by C defined under <stdlib.h> header file to facilitate dynamic memory allocation in C programming. They are:
1. malloc()
2. calloc()
3. free()
4. realloc()
We have used only one of them which is malloc.
malloc(): The “malloc” or “memory allocation” method in C is used to dynamically allocate a single large block of memory with the specified size. It returns a pointer of type void which can be cast into a pointer of any form. It doesn’t Initialize memory at execution time so that it has initialized each block with the default garbage value initially. 
Syntax: ptr = (cast-type*) malloc(byte-size)
For more understandability, this code will help you out.